+++ tests/Fractions.Tests/FractionSpecs/FromString/Method_FromString.cs
@@ -1,5 +1,7 @@
 ﻿using System;
+using System.Collections.Generic;
 using System.Globalization;
+using System.Numerics;
 using FluentAssertions;
 using NUnit.Framework;
 using Tests.Fractions;
@@ -183,5 +185,166 @@ public class When_creating_a_fraction_from_the_string_minus_64_over_40_by_explic
     // German: Das Ergebnis sollte korrekt sein
     public void The_result_should_be_correct() {
         _result.Should().Be(new Fraction(-64, 40));
+    }
+}
+
+[TestFixture]
+public class When_creating_a_fraction_from_a_very_long_decimal_string : Spec {
+    private Fraction _result;
+
+    public override void Act() {
+        _result = Fraction.FromString("123456789987654321.123456789987654321", new CultureInfo("en-US"));
+    }
+
+    [Test]
+    public void The_number_is_parsed_without_any_loss_of_precision() {
+        _result.Should().Be(new Fraction(
+            BigInteger.Parse("123456789987654321123456789987654321"),
+            BigInteger.Pow(10, 18))
+        );
+    }
+}
+
+[TestFixture]
+public class Parsing_a_string_with_scientific_notation_representing_a_large_positive_number : Spec {
+    private Fraction _expectedFraction;
+
+    public override void Arrange() {
+        _expectedFraction = new Fraction(new BigInteger(123456789987654321) * BigInteger.Pow(10, 24));
+    }
+
+    [Test]
+    public void Should_not_work_by_default(
+        [Values("1.23456789987654321E+41", "1.23456789987654321e+41")]
+        string valueToParse) {
+        Invoking(() => Fraction.FromString(valueToParse, new CultureInfo("en-US")))
+            .Should()
+            .Throw<FormatException>();
+    }
+
+    [Test]
+    public void Should_work_without_loss_of_precision_with_NumberStyle_Any(
+        [Values("1.23456789987654321E+41", "1.23456789987654321e+41")]
+        string valueToParse) {
+        Fraction.FromString(valueToParse, NumberStyles.Any, new CultureInfo("en-US"))
+            .Should()
+            .Be(_expectedFraction);
+    }
+
+    [Test]
+    public void Should_not_work_for_decimal_exponents(
+        [Values("1.23456789987654321E+41.5", "1.23456789987654321e+41.5")]
+        string valueToParse) {
+        Invoking(() => Fraction.FromString(valueToParse, NumberStyles.Any, new CultureInfo("en-US")))
+            .Should()
+            .Throw<FormatException>();
+    }
+}
+
+[TestFixture]
+public class Parsing_a_string_with_scientific_notation_representing_a_large_negative_number : Spec {
+    private Fraction _expectedFraction;
+
+    public override void Arrange() {
+        _expectedFraction = new Fraction(new BigInteger(-123456789987654321) * BigInteger.Pow(10, 24));
+    }
+
+    [Test]
+    public void Should_not_work_by_default(
+        [Values("-1.23456789987654321E+41", "-1.23456789987654321e+41")]
+        string valueToParse) {
+        Invoking(() => Fraction.FromString(valueToParse, new CultureInfo("en-US")))
+            .Should()
+            .Throw<FormatException>();
+    }
+
+    [Test]
+    public void Should_work_without_loss_of_precision_with_NumberStyle_Any(
+        [Values("-1.23456789987654321E+41", "-1.23456789987654321e+41")]
+        string valueToParse) {
+        Fraction.FromString(valueToParse, NumberStyles.Any, new CultureInfo("en-US"))
+            .Should()
+            .Be(_expectedFraction);
+    }
+
+    [Test]
+    public void Should_not_work_for_decimal_exponents(
+        [Values("-1.23456789987654321E+41.5", "-1.23456789987654321e+41.5")]
+        string valueToParse) {
+        Invoking(() => Fraction.FromString(valueToParse, NumberStyles.Any, new CultureInfo("en-US")))
+            .Should()
+            .Throw<FormatException>();
+    }
+}
+
+[TestFixture]
+public class Parsing_a_string_with_scientific_notation_representing_a_small_positive_number : Spec {
+    private Fraction _expectedFraction;
+
+    public override void Arrange() {
+        _expectedFraction = new Fraction(new BigInteger(123456789987654321), BigInteger.Pow(10, 17 + 24));
+    }
+
+    [Test]
+    public void Should_not_work_by_default(
+        [Values("1.23456789987654321E-24", "1.23456789987654321e-24")]
+        string valueToParse) {
+        Invoking(() => Fraction.FromString(valueToParse, new CultureInfo("en-US")))
+            .Should()
+            .Throw<FormatException>();
+    }
+
+    [Test]
+    public void Should_work_without_loss_of_precision_with_NumberStyle_Any(
+        [Values("1.23456789987654321E-24", "1.23456789987654321e-24")]
+        string valueToParse) {
+        Fraction.FromString(valueToParse, NumberStyles.Any, new CultureInfo("en-US"))
+            .Should()
+            .Be(_expectedFraction);
+    }
+
+    [Test]
+    public void Should_not_work_for_decimal_exponents(
+        [Values("1.23456789987654321E-24.5", "1.23456789987654321e-24.5")]
+        string valueToParse) {
+        Invoking(() => Fraction.FromString(valueToParse, NumberStyles.Any, new CultureInfo("en-US")))
+            .Should()
+            .Throw<FormatException>();
+    }
+}
+
+[TestFixture]
+public class Parsing_a_string_with_scientific_notation_representing_a_small_negative_number : Spec {
+    private Fraction _expectedFraction;
+
+    public override void Arrange() {
+        _expectedFraction = new Fraction(new BigInteger(-123456789987654321), BigInteger.Pow(10, 17 + 24));
+    }
+
+    [Test]
+    public void Should_not_work_by_default(
+        [Values("-1.23456789987654321E-24", "-1.23456789987654321e-24")]
+        string valueToParse) {
+        Invoking(() => Fraction.FromString(valueToParse, new CultureInfo("en-US")))
+            .Should()
+            .Throw<FormatException>();
+    }
+
+    [Test]
+    public void Should_work_without_loss_of_precision_with_NumberStyle_Any(
+        [Values("-1.23456789987654321E-24", "-1.23456789987654321e-24")]
+        string valueToParse) {
+        Fraction.FromString(valueToParse, NumberStyles.Any, new CultureInfo("en-US"))
+            .Should()
+            .Be(_expectedFraction);
+    }
+
+    [Test]
+    public void Should_not_work_for_decimal_exponents(
+        [Values("-1.23456789987654321E-24.5", "-1.23456789987654321e-24.5")]
+        string valueToParse) {
+        Invoking(() => Fraction.FromString(valueToParse, NumberStyles.Any, new CultureInfo("en-US")))
+            .Should()
+            .Throw<FormatException>();
     }
 }